Moves Books to components
