+++ assets/js/app.js
@@ -21,45 +21,6 @@ class HelloReact extends React.Component {
   }
 }
 
-class Books extends React.Component {
-  constructor() {
-    super();
-    this.state = { books: [] };
-  }
-
-  componentWillMount() {
-    axios
-      .get("http://localhost:4000/api/books")
-      .then(response => {
-        this.setState({ books: response.data.books });
-      })
-      .catch(error => {
-        console.log(error);
-      });
-  }
-
-  render() {
-    const posts = this.state.books.map((book, index) => (
-      <BookListing
-        key={index}
-        id={book.id}
-        title={book.title}
-        description={book.description}
-        genre={book.genre}
-        author={book.author}
-      />
-    ));
-
-    return (
-      <div>
-        <Link to="/">Home</Link>
-        <Link to="/books/new">Add a book</Link>
-        {posts}
-      </div>
-    );
-  }
-}
-
 class Book extends React.Component {
   constructor() {
     super();
